// src/app/models/shop.ts
export interface shopData {
  id: number;
  img: string;
  productTitle: string;
  productName: string;
  desc: string;
  aimg: string;
  amt: number;
  qty: number;
}

export interface cart {
  id1: number;
  img1: string;
  productTitle1: string;
  productName1: string;
  desc1: string;
  aimg1: string;
  amt1: number;
  qty1: number;
}